[PONG:BASE] Refactor: Interfaces extracted to interfaces.ts and IPhysicsObject tweaked.
IPhysicsObject no longer requires bounceY and bounceX methods.
